requirements.txt
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.8
paramiko==3.4.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
scp==0.15.0
urllib3==2.2.2
